Acme-Explorer/src/main/java/services/TripService.java

package services;

import java.text.SimpleDateFormat;
import java.util.ArrayList;
import java.util.Collection;
import java.util.Date;
import java.util.List;
import java.util.Random;
import java.util.concurrent.ThreadLocalRandom;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;
import org.springframework.util.Assert;

import repositories.TripRepository;
import security.LoginService;
import security.UserAccount;
import domain.Admin;
import domain.Application;
import domain.AuditRecord;
import domain.Auditor;
import domain.Category;
import domain.Configuration;
import domain.Explorer;
import domain.HasValue;
import domain.Manager;
import domain.Note;
import domain.Sponsorship;
import domain.Stage;
import domain.Story;
import domain.SurvivalClass;
import domain.Trip;

@Service
@Transactional
public class TripService {

	// Managed repository -----------------------------------------------------

	@Autowired
	private TripRepository			tripRepository;

	// Supporting services ----------------------------------------------------

	@Autowired
	private ConfigurationService	configurationService;

	@Autowired
	private CategoryService			categoryService;

	@Autowired
	private ManagerService			managerService;

	@Autowired
	private ExplorerService			explorerService;

	@Autowired
	private NoteService				noteService;

	@Autowired
	private AuditRecordService		auditRecordService;

	@Autowired
	private AdminService			adminService;

	@Autowired
	private StoryService			storyService;

	@Autowired
	private AuditorService			auditorService;

	@Autowired
	private ApplicationService		applicationService;

	@Autowired
	private SurvivalClassService	survivalClassService;

	@Autowired
	private SponsorshipService		sponsorshipService;

	@Autowired
	private HasValueService			hasValueService;

	@Autowired
	private FinderService			finderService;


	// Constructors -----------------------------------------------------------

	public TripService() {
		super();
	}

	// Simple CRUD methods ----------------------------------------------------

	public Trip create() {

		Manager manager = null;

		Collection<Note> notes = null;
		Collection<AuditRecord> auditRecords = null;
		Collection<Story> stories = null;
		Collection<Application> applications = null;
		Collection<HasValue> hasValues = null;
		Collection<Sponsorship> sponsorships = null;
		Collection<SurvivalClass> survivalClasses = null;
		Collection<Stage> stages = null;
		manager = this.managerService.findByPrincipal();

		stages = new ArrayList<Stage>();
		notes = new ArrayList<Note>();
		auditRecords = new ArrayList<AuditRecord>();
		stories = new ArrayList<Story>();
		applications = new ArrayList<Application>();
		hasValues = new ArrayList<HasValue>();
		sponsorships = new ArrayList<Sponsorship>();
		survivalClasses = new ArrayList<SurvivalClass>();

		final Trip res = new Trip();

		final String ticker = this.generateTicker();

		res.setTicker(ticker);
		res.setNotes(notes);
		res.setAuditRecords(auditRecords);
		res.setStories(stories);
		res.setApplications(applications);
		res.setManager(manager);
		res.setSurvivalClasses(survivalClasses);
		res.setSponsorships(sponsorships);
		res.setStages(stages);
		res.setHasValues(hasValues);
		res.setPrice(0.0);

		return res;

	}

	public Trip save(final Trip trip) {

		final Date actualDate = new Date();

		Assert.notNull(trip);
		Assert.isTrue(trip.getStartDateTrip().before(trip.getEndDateTrip()), "start date before end date");
		Assert.isTrue(trip.getPublicationDate().before(trip.getStartDateTrip()), "publication date before start date");
		Assert.isTrue(trip.getPublicationDate().after(actualDate));
		Assert.isTrue(this.checkPrincipalManager(trip));

		Trip res = null;

		if (trip.getId() == 0) {
			res = this.tripRepository.save(trip);
			res.getManager().getTrips().add(res);
			res.getLegalText().getTrips().add(res);
			res.getRanger().getTrips().add(res);
			if (res.getCategory() != null)
				res.getCategory().getTrips().add(res);

		} else
			res = this.tripRepository.save(trip);

		return res;
	}

	public void removeCategory(final Trip trip) {

		Assert.isTrue(this.checkPrincipalAdmin(trip));

		trip.setCategory(null);
		this.tripRepository.save(trip);
	}

	public Collection<Trip> findAll() {

		Collection<Trip> res = null;
		res = this.tripRepository.findAll();
		return res;
	}

	public Trip findOne(final int tripId) {

		Trip result = null;
		result = this.tripRepository.findOne(tripId);
		return result;
	}

	public Trip findOneToEdit(final int tripId) {

		Trip result = null;
		result = this.tripRepository.findOne(tripId);
		final Manager principal = this.managerService.findByPrincipal();
		Assert.isTrue(result.getManager().equals(principal));
		Assert.isTrue(this.checkPrincipalManager(result) || this.checkPrincipalAuditor(result) || this.checkPrincipalExplorer(result));
		return result;
	}

	public void delete(final Trip trip) {

		Assert.notNull(trip);
		Assert.isTrue(this.checkPrincipalManager(trip));
		Assert.isTrue(this.checkDeleteByPublicationDate(trip));

		this.applicationService.deleteApplications(trip);
		this.auditRecordService.deleteAuditRecords(trip);
		this.storyService.deleteStories(trip);
		this.survivalClassService.deleteSurvivalClasses(trip);
		this.finderService.deleteReferenceTrip(trip);
		this.hasValueService.deleteByTrip(trip);
		this.sponsorshipService.deleteSponsorShips(trip);
		this.noteService.deleteNotes(trip);

		this.tripRepository.delete(trip);

	}
	// Other business methods -------------------------------------------------

	public boolean canEdit(final Trip trip) {

		Assert.notNull(trip);

		boolean result = false;
		final Date date = new Date();

		final Manager manager = this.managerService.findByPrincipal();
		if (manager != null && manager.equals(trip.getManager()) && trip.getPublicationDate().after(date))
			result = true;

		return result;
	}

	public boolean checkPrincipalManager(final Trip trip) {

		boolean res = false;

		final Manager manager = this.managerService.findByPrincipal();
		if (manager != null && manager.equals(trip.getManager()))
			res = true;
		return res;
	}

	public boolean checkPrincipalExplorer(final Trip trip) {

		boolean res = false;

		final Collection<Explorer> explorerers = this.tripRepository.explorerPerTrip(trip.getId());
		final Explorer login = this.explorerService.findByPrincipal();
		if (explorerers.contains(login))
			res = true;
		return res;

	}
	public boolean checkPrincipalAuditor(final Trip trip) {

		boolean res = false;

		final Collection<Auditor> auditors = this.tripRepository.auditorPerTrip(trip.getId());
		final Auditor login = this.auditorService.findByPrincipal();
		if (auditors.contains(login))
			res = true;
		return res;
	}

	public boolean checkPrincipalAdmin(final Trip trip) {

		final boolean res = false;

		final Admin login = this.adminService.findByPrincipal();

		Assert.notNull(login);
		return res;
	}

	public Collection<Trip> findByManager(final Manager manager) {

		Collection<Trip> result = null;
		result = this.tripRepository.listTripPerManager(manager.getId());
		return result;
	}

	public boolean checkDeleteByPublicationDate(final Trip trip) {

		final Date date = new Date();
		return trip.getStartDateTrip().after(date);

	}

	public Manager findByPrincipal() {

		Manager result = null;
		final UserAccount userAccount = LoginService.getPrincipal();
		result = this.findByUserAccountId(userAccount.getId());
		return result;
	}

	public Manager findByUserAccountId(final int userAccountId) {

		Manager result = null;
		result = this.managerService.findByUserAccountId(userAccountId);
		return result;
	}

	public Collection<Trip> listTripPerManager(final int managerId) {

		Collection<Trip> res = null;
		res = this.tripRepository.listTripPerManager(managerId);
		return res;
	}

	public Collection<Trip> listTripPerExplorer(final Explorer explorer) {

		Collection<Trip> res = null;
		res = this.tripRepository.listTripPerExplorer(explorer);
		return res;
	}

	public Collection<Double> avgMinMaxDevApplicationsPerTrip() {
		Collection<Double> res = null;
		res = this.tripRepository.avgMinMaxDevApplicationsPerTrip();
		return res;
	}

	public Collection<Double> avgMinMaxDevPriceOfTheTrips() {
		Collection<Double> res = null;
		res = this.tripRepository.avgMinMaxDevPriceOfTheTrips();
		return res;
	}

	public String ratioOfTripsCancelledVsTotalTripsOrganized() {
		String res = null;
		res = this.tripRepository.ratioOfTripsCancelledVsTotalTripsOrganized();
		return res;
	}

	public Collection<Trip> listingTrips10PercentMoreApplicantionsThanAvg() {
		Collection<Trip> res = null;
		res = this.tripRepository.listingTrips10PercentMoreApplicantionsThanAvg();
		return res;
	}

	public Collection<Double> minMaxAvgDevNotesPerTrip() {
		Collection<Double> res = null;
		res = this.tripRepository.minMaxAvgDevNotesPerTrip();
		return res;
	}

	public Collection<Double> minMaxAvgDevAuditRecordPerTrip() {
		Collection<Double> res = null;
		res = this.tripRepository.minMaxAvgDevAuditRecordPerTrip();
		return res;
	}

	public Collection<Auditor> auditorPerTrip(final int tripId) {
		Collection<Auditor> res = null;
		res = this.tripRepository.auditorPerTrip(tripId);
		return res;
	}

	public String ratioOfTripsWithAnyAuditRecord() {
		String res = null;
		res = this.tripRepository.ratioOfTripsWithAnyAuditRecord();
		return res;
	}

	public Collection<Trip> findAvailableTrips(final Date moment) {
		Assert.notNull(moment);
		Collection<Trip> result = null;
		result = this.tripRepository.findAvailableTrips(moment);
		return result;
	}

	public String generateTicker() {

		String result;
		Trip trip;
		final Random random = new Random();
		final String letters = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";

		while (true) {
			result = new SimpleDateFormat("yyMMdd").format(new Date()) + "-";
			for (int i = 0; i <= 3; i++) {
				final int index = (int) (random.nextFloat() * letters.length());
				result += letters.charAt(index);
			}
			trip = this.tripRepository.findByTicker(result);
			if (trip == null)
				break;
		}

		return result;
	}

	public Double getPrice(final Trip trip) {
		double price = 0.0;
		final Configuration configuration = this.configurationService.findAll().iterator().next();
		final double vat = configuration.getVat() / 100;
		double aux = 0.0;
		double result = 0.0;

		for (final Stage s : trip.getStages()) {
			price = s.getPrice();
			aux = price * vat;
			result = price + aux;

		}
		final double res = Math.round((result) * 100.0) / 100.0;
		trip.setPrice(res);

		return result;
	}
	public Sponsorship getRandomSponsorship(final Trip trip) {
		Sponsorship res = null;
		final Collection<Sponsorship> sponsorships = this.sponsorshipService.findAll();
		int index = 0;

		final List<Sponsorship> listSponsorships = new ArrayList<Sponsorship>(sponsorships);
		index = ThreadLocalRandom.current().nextInt(0, sponsorships.size());

		res = listSponsorships.get(index);
		return res;
	}

	public Collection<Trip> findTripPerPublicationDate() {
		final Date actualDate = new Date();
		Collection<Trip> trips = new ArrayList<Trip>();

		trips = this.tripRepository.findTripPerPublicationDate(actualDate);

		return trips;
	}

	public void cancelTrip(final Trip trip, final String cancelReason) {

		final Date actualDate = new Date();

		Assert.isTrue(trip.getStartDateTrip().after(actualDate));

		trip.setCancelledReason(cancelReason);
		trip.setPublicationDate(actualDate);

		this.tripRepository.save(trip);
	}

	public Collection<Trip> findTripPerKeywordR(final String keyword) {
		Collection<Trip> trips = null;
		trips = new ArrayList<Trip>();
		String aux = "Trip";
		final Date moment = new Date();

		if (keyword == null)
			trips = this.findTripPerPublicationDate();
		if (keyword != null) {
			aux = keyword;
			trips = this.tripRepository.findTripPerKeyword(aux, moment);
		}

		return trips;

	}

	public Collection<Trip> findTripPerCategory(final Integer categoryId) {
		Collection<Trip> trips = null;
		trips = new ArrayList<Trip>();
		Category category;
		final Date moment = new Date();

		if (categoryId == null)
			trips = this.findTripPerPublicationDate();
		if (categoryId != null) {
			category = this.categoryService.findOne(categoryId);
			trips = this.tripRepository.findTripPerCategory(category, moment);
		}

		return trips;
	}

	public Boolean hasThisTripAnyApplicationFromThisExplorer(final int tripId, final int explorerId) {
		return this.tripRepository.hasThisTripAnyApplicationFromThisExplorer(tripId, explorerId);
	}
}
